LayoutTests/media/audio-constructor-src-expected.txt

Test that Audio("url") constructor loads the specified resource.

TEST(audio instanceof HTMLAudioElement) OK
EVENT(loadstart) TEST(relativeURL(audio.currentSrc)=='content/test.wav') OK
EVENT(load)
END OF TEST
